Change in DataInit, move file to initialization directory
